requirements.txt
langchain
langchain-community
openai
faiss-cpu
tiktoken
pandas
numpy
python-dotenv